Refactored identical code for accepting/rejecting booking requests to sharedActions.ts
