+++ src/routes/(app)/booking/admin/[id]/+page.server.ts
@@ -6,9 +6,7 @@ import { zod } from "sveltekit-superforms/adapters";
 import { bookingSchema } from "../../schema";
 import { authorize } from "$lib/utils/authorization";
 import apiNames from "$lib/utils/apiNames";
-import type { Actions } from "./$types";
-import sendNotification from "$lib/utils/notifications";
-import { NotificationType } from "$lib/utils/notifications/types";
+import { actions } from "../sharedActions";
 
 export const load = async ({ locals, params }) => {
   const { prisma, user } = locals;
@@ -68,81 +66,4 @@ export const load = async ({ locals, params }) => {
   };
 };
 
-export const actions: Actions = {
-  accept: async (event) => {
-    const { request, locals } = event;
-    const { prisma, user } = locals;
-    const formData = await request.formData();
-    const id = formData.get("id");
-
-    if (id && typeof id === "string") {
-      await prisma.bookingRequest.update({
-        where: {
-          id,
-        },
-        data: {
-          status: "ACCEPTED",
-        },
-      });
-
-      const request = await prisma.bookingRequest.findFirst({
-        where: {
-          id,
-        },
-        select: {
-          bookerId: true,
-          event: true,
-        },
-      });
-
-      if (request && request.bookerId != null && user && user.memberId) {
-        sendNotification({
-          title: "Booking request accepted",
-          message: `Your booking request for ${request.event} has been accepted`,
-          type: NotificationType.BOOKING_REQUEST,
-          link: `/booking`,
-          memberIds: [request.bookerId],
-          fromMemberId: user.memberId,
-        });
-      }
-    }
-  },
-  reject: async (event) => {
-    const { request, locals } = event;
-    const { prisma, user } = locals;
-    const formData = await request.formData();
-    const id = formData.get("id");
-
-    if (id && typeof id === "string") {
-      await prisma.bookingRequest.update({
-        where: {
-          id,
-        },
-        data: {
-          status: "DENIED",
-        },
-      });
-
-      const request = await prisma.bookingRequest.findFirst({
-        where: {
-          id,
-        },
-        select: {
-          bookerId: true,
-          event: true,
-        },
-      });
-
-      if (request && request.bookerId != null && user && user.memberId) {
-        sendNotification({
-          title: "Booking request denied",
-          message: `Your booking request for ${request.event} has been denied`,
-          type: NotificationType.BOOKING_REQUEST,
-          link: `/booking`,
-          memberIds: [request.bookerId],
-          fromMemberId: user.memberId,
-        });
-      }
-    }
-  },
-};
+export { actions };
